+++ Nos/Service/NostrIdentifier/NostrIdentifier.swift
@@ -14,6 +14,9 @@ enum NostrIdentifier {
     /// A nostr public key
     case npub(publicKey: RawAuthorID)
 
+    /// A nostr private key
+    case nsec(privateKey: String)
+
     /// A nostr note
     case note(eventID: RawEventID)
 
@@ -24,7 +27,7 @@ enum NostrIdentifier {
     case nevent(eventID: RawEventID, relays: [String], eventPublicKey: String?, kind: UInt32?)
 
     /// A nostr replaceable event coordinate
-    case naddr(eventID: RawEventID, relays: [String], eventPublicKey: String, kind: UInt32)
+    case naddr(replaceableID: RawReplaceableID, relays: [String], authorID: RawAuthorID, kind: UInt32)
 
     /// A nostr address
 
@@ -34,15 +37,17 @@ enum NostrIdentifier {
     static func decode(bech32String: String) throws -> NostrIdentifier {
         let (humanReadablePart, data) = try Bech32.decode(bech32String)
         switch humanReadablePart {
-        case Nostr.publicKeyPrefix:
+        case NostrIdentifierPrefix.publicKey:
             return try decodeNostrPublicKey(data: data)
-        case Nostr.notePrefix:
+        case NostrIdentifierPrefix.privateKey:
+            return try decodeNostrPrivateKey(data: data)
+        case NostrIdentifierPrefix.note:
             return try decodeNostrNote(data: data)
-        case Nostr.profilePrefix:
+        case NostrIdentifierPrefix.profile:
             return try decodeNostrProfile(data: data)
-        case Nostr.eventPrefix:
+        case NostrIdentifierPrefix.event:
             return try decodeNostrEvent(data: data)
-        case Nostr.addressPrefix:
+        case NostrIdentifierPrefix.address:
             return try decodeNostrAddress(data: data)
         default:
             throw NostrIdentifierError.unknownPrefix
@@ -57,6 +62,16 @@ enum NostrIdentifier {
             throw NostrIdentifierError.unknownFormat
         }
         return .npub(publicKey: publicKey)
+    }
+
+    /// Decodes nsec data into a `NostrIdentifier.nsec`.
+    /// - Parameter data: The encoded nsec data.
+    /// - Returns: The `.nsec` with the private key.
+    private static func decodeNostrPrivateKey(data: Data) throws -> NostrIdentifier {
+        guard let privateKey = SHA256Key.decode(base5: data) else {
+            throw NostrIdentifierError.unknownFormat
+        }
+        return .nsec(privateKey: privateKey)
     }
 
     /// Decodes npub data into a `NostrIdentifier.note`.
@@ -78,11 +93,13 @@ enum NostrIdentifier {
         var publicKey = ""
         var relays: [String] = []
         for element in tlvElements {
-            switch element {
-            case .special(let value):
-                publicKey = value
-            case .relay(let value):
-                relays.append(value)
+            switch element.type {
+            case .special:
+                publicKey = SHA256Key.decode(base8: element.value)
+            case .relay:
+                if let string = String(data: element.value, encoding: .ascii) {
+                    relays.append(string)
+                }
             default:
                 break
             }
@@ -102,17 +119,17 @@ enum NostrIdentifier {
         var publicKey: String?
         var kind: UInt32?
         for element in tlvElements {
-            switch element {
-            case .special(let value):
-                eventID = value
-            case .relay(let value):
-                relays.append(value)
-            case .author(let value):
-                publicKey = value
-            case .kind(let value):
-                kind = value
-            case .unknown:
-                break
+            switch element.type {
+            case .special:
+                eventID = SHA256Key.decode(base8: element.value)
+            case .relay:
+                if let string = String(data: element.value, encoding: .ascii) {
+                    relays.append(string)
+                }
+            case .author:
+                publicKey = SHA256Key.decode(base8: element.value)
+            case .kind:
+                kind = UInt32(bigEndian: element.value.withUnsafeBytes { $0.load(as: UInt32.self) })
             }
         }
 
@@ -125,25 +142,27 @@ enum NostrIdentifier {
     private static func decodeNostrAddress(data: Data) throws -> NostrIdentifier {
         let tlvElements = TLVElement.decodeElements(data: data)
 
-        var eventID = ""
+        var replaceableID = ""
         var relays: [String] = []
-        var publicKey: String = ""
+        var authorID: String = ""
         var kind = UInt32.max
         for element in tlvElements {
-            switch element {
-            case .special(let value):
-                eventID = value
-            case .relay(let value):
-                relays.append(value)
-            case .author(let value):
-                publicKey = value
-            case .kind(let value):
-                kind = value
-            case .unknown:
-                break
+            switch element.type {
+            case .special:
+                if let string = String(data: element.value, encoding: .ascii) {
+                    replaceableID = string
+                }
+            case .relay:
+                if let valueString = String(data: element.value, encoding: .ascii) {
+                    relays.append(valueString)
+                }
+            case .author:
+                authorID = SHA256Key.decode(base8: element.value)
+            case .kind:
+                kind = UInt32(bigEndian: element.value.withUnsafeBytes { $0.load(as: UInt32.self) })
             }
         }
 
-        return .naddr(eventID: eventID, relays: relays, eventPublicKey: publicKey, kind: kind)
+        return .naddr(replaceableID: replaceableID, relays: relays, authorID: authorID, kind: kind)
     }
 }