+++ NosTests/Models/ReportTests.swift
@@ -1,0 +1,11 @@
+import XCTest
+
+final class ReportTests: XCTestCase {
+
+    func testFindCategory() throws {
+        XCTAssertEqual(
+            ReportCategory.findCategory(from: "PN"),
+            ReportCategoryType.allCategories[8]
+        )
+    }
+}