+++ CHANGELOG.md
@@ -8,13 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Added a filter button to the Home tab that lets you browse all notes on a specific relay.
+- Fix a bug where multiple connections could be opened with the same relay.
+- Fixed an issue where Profile views would sometimes not display any notes.
+- Add impersonation flag category and better NIP-56 mapping.
 - Add a Tap to Refresh button in empty profiles.
-- Update the reply count shown below each Note in a Feed.
+- Support nostr:naddr links to text and long-form content notes.
+- Update the reply count shown below each note in a Feed.
 - Removed follower count from profile screen.
 - Fixed deep linking to profiles and notes.
 - Fixed issue where some nostr:nprofile references did not appear as links.
 - Decode nprofile, nevent, and naddr NIP-19 entities.
+- Refactor away direct usages of Bech32 and TLV and replace with NostrIdentifier (which still uses both).
 
 ## [0.1.20] - 2024-07-10Z